+++ install/SCRIPTS/inst_games.sh
@@ -27,6 +27,7 @@ SPIN='/-\|'
 INSTALL_DIR=$(gnustep-config --variable=GNUSTEP_LOCAL_APPS)
 TEMPFILE=$(mktemp /tmp/agno-XXXXX)
 trap "rm -f $TEMPFILE" EXIT
+SLEEP=2
 
 ### End of VARS
 ################################################