+++ install/SCRIPTS/spinner.sh
@@ -45,3 +45,27 @@ done
 ### End of Spinner
 #############################################
 
+
+#############################################
+### Timer
+### set: DELAY=x
+### default: DELAY=5
+### MSG="a message"
+function timer
+{
+if [ $DELAY ];then
+	i=$DELAY
+else
+	i=5
+fi
+
+printf "$MSG "
+while [ $i -gt 0 ]
+do
+	sleep 1
+	let i--
+	printf "\b$i"
+done
+printf "\nGo!\n"
+
+}